ISAICP-2369: Update the selectors of the navigation menu.
These were inheriting the generic selectors for the sidebar menu but these are
not working well for the navigation menu. Ilias mentioned that this was never
the intention, and this step was even commented out at some point in the develop
branch to prevent failures.

This means that this will need to be re-themed.