ISAICP-2603: :(.. Rewrite process results. The idea here is that there was a bug with the request graphs. The priority of the request graphs was not respected. What changed is that the entity graphs are loaded for the entity type and then are filtered for each priority graph. The issue is that we have the request graph ids but we get the results indexed by graph uris. To fill the gap, we are loading the full list of the graphs.
WARNING: This might cause issues if multiple graphs for multiple bundles are defining the same uri.
E.g. If a bundle 1 has a default graph http://uri1 and a draft graph http://uri2
and a bundle 2 has a default graph http://uri2 and a draft graph http://uri3
and the graph draft is requested for an entity, if it is of a type bundle 1 then the wrong graph will
be loaded.
This is caused because the query must be bundle agnostic for the cases that a bundle is not provided
so all bundles are taken into account.