ISAICP-2369: Remove workaround that attempted to delete a solution twice
This has been fixed in the meantime now that the doDelete() method correctly
deletes from all graphs. This went unnoticed because OG was not clearing its
static caches after an entity was deleted.

Now that we have fixed static cache clearing in OG this workaround has
surfaced. We can safely remove it.