ISAICP-4086: Uninstalling default config after installing a module is useless at the moment.
When we deploy new functionality the following sequence takes place:
1. New code is deployed.
2. Update hooks are executed, this includes the post-update hooks.
3. Configuration is synced.

When we enable the Message Subscribe module in the post-update hook, the
config_sync module will snapshot the current default config of the
module. At this point the unneeded default config will be part of the
snapshot. We then delete the default config, but this will be restored
during the configuration synchronization step, because config_sync
currently does not support the deletion of existing configuration.

Ref. https://www.drupal.org/project/config_sync/issues/2914536

We are going to let this slide for the moment since the default config
supplied by the Message Subscribe module is disabled and harmless.