ISAICP-4111: Turn the entity message helper into a basic service instead of a base class.
The use cases for messages are too diverse to use a base class or a
trait. It is not feasible to create an interface that is valid for all
possible use cases. If this is a service we can still reuse the code
without having to define a unified interface.