ISAICP-5091: Restore build properties for Solr cores.
The build properties for the Solr server was integrated in the properties for
the Solr cores when we provided Docker integration. This made it possible to
host each core on a different Solr server. Unfortunately the properties were
not applied on all existing Phing targets which broke the Solr integration on
local (non-Docker) environments.