ISAICP-4974: We are switching from the open source Matomo solution to the proprietary Webtools Analytics solution for our visitor tracking.
We cannot install Webtools Analytics on our test and development instances and
no direct reporting will happen any more to Matomo, so it is pointless to
install this software on any of our enviroments. The only thing we can do to
test is to verify that the proprietary JavaScript is present in the page.
Remove all targets related to managing a local Matomo instance.