ISAICP-4974: Restore the Matomo settings in the original location.
Since this is configuration which is already exported by the Matomo module we
cannot export this in another module since core will complain that the config
already exists. Once we split Joinup in the distro part and the ISA2 part then
this file should be moved to the export folder.