ISAICP-5147: Attempt to work around limitations in the SPDX module architecture.
The SPDX module has defined the SPDX licences as a bundle of the RDF entity type
defined in the RDF Entity module. This means they cannot use base fields for
their basic properties and have to resort to "normal" Field API fields exported
in config. We need to override this config, but we cannot do it in the
joinup_licence module since core will rightfully complain that the config
already exists in active storage. So we need to put it in the profile.