ISAICP-5620: Do not clear the cache until after updates are performed.
Since an update might introduct new modules with new service dependencies we
should first perform all updates before clearing the cache. The cache clear
command bootstraps the DrupalKernel but this might crash because it will try
to populate the dependency injection container with services that depend on
services defined in the new modules which are not enabled at this point.

@see https://www.drupal.org/node/3067480