ISAICP-4980: Make sure to update content through the API rather than the UI.
This solves a difficult caching problem that cannot be rectified by our standard
solutions like @clearStaticCache.

Messages have an entity reference to a node, but if we create the node through
the API and then later publish it through the UI then the API thread still has
a reference to the unpublished version in its static cache.