ISAICP-6098: Index the new Entity Metadata fields that now host the 'pinned in' data.
These replace the previous 'sticky' and 'field_is_pinned_in' fields. This data
is available for both nodes and RDF entities so it would be natural to index
them as a single aggregated field, but this is not supported for computed fields
in the current version of Search API. So for now they are indexed separately.