ISAICP-6098: The statistics module should only create its own meta entities.
Before the introduction of the meta entities for content that is pinned in a
group the only meta entities that were used in Joinup were for tracking
statistics. The joinup_stats module naively treated all meta entities as
being used for statistics.

This commit ensures that the module only creates meta entities for its own
purpose.