ISAICP-6099: Move code related to featuring content site wide into a separate module.
This was located in the joinup_front_page module, but this was not correct
since content is featured site wide rather than just on the front page.

Future refactoring will also introduce an interface and trait to use in
bundle classes so it is a good idea to make this a small standalone module.