ISAICP-6259: Eliminate dependency on joinup_core.
The EntityBundlePairItem field type is only used by the subscription module. It
was probably placed in joinup_core to promote reuse, but this introduces a
gnarly dependency chain.

If this is eligible for reuse in another Joinup subsystem we can opt to put this
in a dedicated module, but for now this is not yet needed.