ISAICP-6259: Move event subscriber back to joinup_subscription.
Having this event subscriber in joinup_subscription ensures that we have no hard
module dependencies between the two modules. Since the event is fired in the
joinup_notification module, we can decouple them by placing the subscriber here.
If joinup_notification is disabled, then the event never fires and the code is
inactive. If joinup_subscription is disabled, the subscriber doesn't respond.