ISAICP-6259: There is a difference in handling of messages with a single recipient and multiple recipient.
My earlier refactoring missed the fact that a message intended for multiple
users is saved using the anonymous user as the owner, and uses the mail sending
mechanism of the Message Notify module to send them.

This commit restores this functionality while still keeping the cleaned up
methods, and adds additional documentation to each method explaining the use
cases.