ISAICP-6348: Do not use the creation date of the initial draft of a document as the canonical publication date.
This field is intended to allow editors to override the publication date of
official documents to match the actual date the content was initially released
to the public, regardless of when it was published on Joinup. Setting this
to the initial creation date of the first draft on Joinup defeats this
purpose.