ISAICP-6348: Work around a premature cache rebuild performed during drush deploy.
Drush will rebuild the cache just prior to importing configuration during a
deployment. As part of this the entities are rebuilt. We are introducing a new
module with an interface that will be used in some of our existing entities, but
this new module will only be enabled as part of the configuration update. When
drush clears the cache the module is not yet enabled and the interface is not
found since it is autoloaded dynamically.

The workaround simply adds the missing classes to the classmap so they are also
available even if the module is not enabled yet.