ISAICP-6399: Keep track of whether an RDF entity has been previously published.
This hack provides us the information we need, until we have revisionable RDF
entities. This was originally in the joinup_group module but we now also need
it for non-group content such as asset releases and distributions. Since this
hack is only used for notifications it has been moved to joinup_notification.