ISAICP-6475: Recreate the In the spotlight view from scratch.
The first iteration was showing a view of entity subqueues and this initially
seemed to be working well, but it turned out to be impossible to number the
content as the entire entity queue was rendered in 1 single Views row.
This new version is a "standard" View of nodes, with a relationship to the
entity queue table, and using the queue position as a sort parameter.