ISAICP-6563: Fix previously unintentionally broken test.
Previously the editor format of the abstract field was not correctly set, and it
was not possible for a moderator to edit the abstract field. It would still
contain the original abstract text instead of the field being cleared. Now it is
cleared correctly and the abstract text no longer appears in the page.