ISAICP-6595: Find a good balance between 'finger' and 'mouse' swiping.
On screen sizes typical for mobile devices we always scroll the number of
visible items. This gives a good balance between the feeling of inertia present
on touch enabled devices, while still feeling natural when using the mouse on a
desktop device.

On the largest screen size we are scrolling to the actual swiped content item,
because in this case 5 items are shown out of a total of 12. If we would try to
scroll 5 items this becomes an unnatural, because scrolling the third time
would only scroll items 11 + 12 into view, which is a very short scroll
distance.

This seems to be a limitation or possibly a configuration problem in the Slick
JS library. Even though endless scrolling is enabled, when reaching the end of
the list it will initiate a scroll to the last (12th) item, instead of
respecting the configured scroll distance, and scrolling past the repeat point.

Scrolling 2 items on a very wide screen feels abrupt and unnatural. Possible
solutions are to either show 6 items, fix the looping problem in Slick JS, or
scrolling to the actual swiped content item. The last option is the easiest
solution, but has the drawback that it doesn't feel fluid when finger-swiping
on a touch enabled device. This is acceptable though since this screen size
is very uncommon on mobile devices.